change floats to integers with a 10^-6 precision for our sensors, much more usable on uC than floats

+++ ANoiseSensor.hpp
@@ -13,7 +13,7 @@ namespace athome {
             ~ANoiseSensor();
 
             virtual uint8_t *getSample() = 0;
-            virtual float   getLastSample() const = 0;
+            virtual int32_t getLastSample() const = 0;
             ISensorScale    getEstimate();
         };
     }
